CoachNote#author should link the user not the faculty
