+++ app/graphql/types/coach_note_type.rb
@@ -1,7 +1,7 @@
 module Types
   class CoachNoteType < Types::BaseObject
     field :id, ID, null: false
-    field :author, Types::CoachType, null: true
+    field :author, Types::NoteAuthorType, null: true
     field :note, String, null: false
     field :created_at, String, null: false
   end
