+++ app/graphql/types/note_author_type.rb
@@ -1,0 +1,14 @@
+module Types
+  class NoteAuthorType < Types::BaseObject
+    field :id, ID, null: false
+    field :name, String, null: false
+    field :title, String, null: false
+    field :avatar_url, String, null: true
+
+    def avatar_url
+      if object.avatar.attached?
+        Rails.application.routes.url_helpers.rails_representation_path(object.avatar_variant(:thumb), only_path: true)
+      end
+    end
+  end
+end
